collector/requirements.txt
requests==2.32.3
psycopg2-binary==2.9.9
sqlalchemy==2.0.36
anthropic
google-genai
